+++ model/mmwave-ue-phy.h
@@ -115,7 +115,7 @@ public:
 
   void SetPhyMacConfig (Ptr<MmWavePhyMacCommon> config);
 
-  virtual AntennaArrayModel::BeamId GetBeamId (uint8_t rnti) const override
+  virtual AntennaArrayModel::BeamId GetBeamId (uint16_t rnti) const override
   {
     NS_UNUSED (rnti);
     NS_FATAL_ERROR ("ERROR");
@@ -185,9 +185,6 @@ private:
   Time m_wbCqiPeriod;       /**< Wideband Periodic CQI: 2, 5, 10, 16, 20, 32, 40, 64, 80 or 160 ms */
   Time m_wbCqiLast;
 
-  VarTtiAllocInfo::TddMode m_prevSlotDir;
-
-  SlotAllocInfo m_currSlotAllocInfo;
   std::vector< std::list<TbAllocInfo> > m_ulTbAllocQueue;       // for storing info on future UL TB transmissions
   bool m_ulGrant;               // true if no uplink grant in subframe, need to transmit UL control in PUCCH instead
   bool m_slotAllocInfoUpdated;